requirements.txt
# Core MCP server
fastmcp==3.2.4
mcp==1.27.0

# Web framework (frontend API)
Flask==3.1.3
flask-cors==6.0.2

# Desktop UI
pywebview==6.2

# Vector database client
qdrant-client==1.17.1

# Embedding model
sentence-transformers==5.4.1
transformers==5.5.4
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.10.2

# PyTorch - GPU (CUDA 12.1)
# Install separately:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
# For CPU only:
#   pip install torch
torch==2.5.1+cu121

# Settings & validation
pydantic==2.13.0
pydantic-settings==2.13.1

# Utilities
numpy==2.4.4
scikit-learn==1.8.0
scipy==1.17.1
tqdm==4.67.3
python-dotenv==1.2.2
